Adding support for HID interface USB ZWave controllers (ThinkStick, Wayne-Dalton branded, etc).  This is the initial implementation, there are bugs to be fixed and further analysis/tweaks will be needed to make it stable and reduce packet errors.
Currently the most compatible  firmware is ThinkStick v4.53; it emits several errors during initialization but afterwards seems to communicate normally with other ZWave devices.

HID support by default is not enabled in the MinOZW sample application.  You will need to swap the commented call to Manager::Get()->AddDriver() line in MinOZW main() so that it uses the HID driver instead of the serial one.