Finally get MultiInstance support functional. There were a number of details to attend to. The largest was discovering that the non-instance value, that is the value that isn't encapsulated in the MultiInstance class, is required to be used. Other devices replicate the non-instance value with an end point value. Since different devices can replicate the non-instance value with different endpoints, it is easier to just not use the non-instance value.
The MultiInstance class now sets up a mapping for the two cases. The xml
configuration files can set the mapping using the mapping= attribute. Its
values can be "all" or "endpoints". The "all" setting is the default and
uses the non-instance value and all end points as values. The "endpoints"
setting ignores the non-instance value and just uses the endpoints. The
default of treating the non-instance value as endpoint 1 and starting with
endpoint 2 is still in the code and can be set up to be used but doesn't
have an xml option as yet.

Some changes to the Msg object to differentiate the instance to end point
mapping were necessary. A flags field was added to handle the different
encapsulation options.

The Frostdale device under-reported its endpoint count so an endpoint=
attribute was added.

MultiInstance no longer tries to use the Find operations as devices don't
seem to support it very well if at all. Now the code will send a single
capability get if all classes are the same or a capability get per end
point otherwise.

The find code still exists and a findsupport= attribute has been added in
case we want to wire it back into the xml configuration. It currently isn't
connected in.

Add support for WakeUp V2 but it isn't complete. Just message formatting
but discovery still needs to be changed to capture the data.

Remove the MultiEndPointFindSupported xml configuration option as it is no
longer necessary.