Reverted and resetted everything and added chenged files
All old updates:

*Update .Net wrapper for new Manager API's:
Added Methods to ZWManager corresponding to the new API methods in
Manager.
ToDo: Change the exampes so they use the new API methods (Use
Notification::Type_ControllerCommand as the reply for the API methods.
The (OpenZWave::Driver::ControllerState)_notification->GetEvent(),
(OpenZWave::Driver::ControllerError)_notification->GetNotification() can
be used as parameters for the original callback method, I believe)!

*Edited .Net wrapper and OZWForm to use the new methods:
Edited the .Net wrapper and OZWForm to work with the new methods. For
now I have used only one way of adding nodes, which always uses the
security enabled way. I have no other devices, so can't test if this is
right or not. Furthermore, I split the implementation and definition of
the methods between the ZWManager.cpp and ZWManager.h. But the methods
are just one line calls to the manager, and as some other methods like
that in ZWManager are just implemented in the header file, this might be
part of the coding standard? If so, I'll have to change that.

When BeginControllerCommand is fully deprecated everything should be
looked over and edited further. For now my changes seem to work as they
should, but I have had some problems with assertions failing in
"output.c", when I was trying to remove a node, but it does not seem to
fail every time. I believe this is not due to my code, so I let it be.

Finally, my code should be tested more extensively, by someone with more
devices and (maybe) more experience with the project, to test whether
everything actually works as it should!

*Updated OZWForm with a new menu item to add secure devices