Add Devolo Home Control Radiator Thermostat
The Devolo Home Control Radiator Thermostat is identical to the Danfoss Z Thermostat. It's rumored that there are some firmware differences, but it is functionally 100% identical to the Danfoss model, works fine with the danfoss/z.xml and has even the Danfoss manufacturer ID.
So it would be nice to add the configuration to the config file, it's just the one line with the different product id.