Set param 83/84 as read only
Parameter 83 is a reference to the Color Switch V1 Command Class which
should be used to change its color instead of this parameter setting.

The Multilevel Switch Command class can be used to dim the LED strip
also using Night Light Mode.

These are both command classes that are used to change colors and the
dimness of the Nightlight LED.