possible fix (or workaround) for a deadlock
My OpenZwave stopped responding multiple times a day. Looking into this I found out that the driverThread was hanging while trying to get a lock for m_nodeMutex.

2017-01-20 10:54:57.871 Warning, CheckCompletedNodeQueries m_allNodesQueried=0 m_awakeNodesQueried=1
2017-01-20 10:54:57.874 Always,       Locking 827319692 from void OpenZWave::Driver::CheckCompletedNodeQueries()
2017-01-20 10:54:57.876 Always,       Locking 818914372 from bool OpenZWave::Manager::IsNodeFailed(uint32, uint8)
2017-01-20 10:54:57.878 Always,       Locking 827319692 from void OpenZWave::Driver::CheckCompletedNodeQueries() ... done with ret: true
2017-01-20 10:54:57.882 Always,       Unlocking 827319692
2017-01-20 10:54:57.884 Always,       Locking 818914372 from bool OpenZWave::Manager::IsNodeFailed(uint32, uint8) ... done with ret: true
2017-01-20 10:54:57.891 Always,       LockGuard destructor 818914372, isSignaled? false
2017-01-20 10:54:57.898 Always,       Locking 818914264 from bool OpenZWave::Manager::RefreshValue(const OpenZWave::ValueID&)
2017-01-20 10:54:57.900 Always,       Locking 818914264 from bool OpenZWave::Manager::RefreshValue(const OpenZWave::ValueID&) ... done with ret: true
2017-01-20 10:54:57.904 Warning, CheckCompletedNodeQueries all=0, deadFound=1 sleepingOnly=1
2017-01-20 10:54:57.906 Always,       LockGuard destructor 827319692, isSignaled? false
2017-01-20 10:54:57.909 Always,       LockGuard destructor 818914264, isSignaled? false
2017-01-20 10:54:57.925 Error, MutexImpl::UnLock failed with error: 2 (1)

In this log you see that 827319692 is trying to unlock twice, once with a call to the LockGuard::Unlock() and once via the destructor. But in the meantime another thread (818914264) got the lock and also tries to unlock the mutex. So now 1 unlock fails and as a result the m_lockcount is now -1. The next lock increases this again to 0, but as isSignaled() returns true when m_lockcount is 0 the destructor of the LockGuard does not unlock this, and thus the mutex is locked forever.
To work around this I removed all manual calls to the LockGuard::Unlock() method and instead made a scope around them, now no LockGuard should try to unlock twice.
I did not look at the MutexImplementations for windows and winRT.
Does this make any sense?

What i am not sure, in case the unlocking fails, would it help to increase the m_lockCount again?