Fix ValueList::SetByValue and add ValueList::SetByIndex()
continuing from
https://github.com/OpenZWave/open-zwave/pull/1035
and
https://github.com/OpenZWave/open-zwave/issues/1065

When using the
```
bool Manager::SetConfigParam
(
		uint32 const _homeId,
		uint8 const _nodeId,
		uint8 const _param,
		int32 _value,
		uint8 const _size
)
```
function, this clearly takes a value as a argument.
Comming down this path, the ValueList::SetByValue need to get a index for the value. That's why I made the first PR.
Now that this got reverted i again got crashes.
So this commit adds a SetByIndex() function. Now both pathes should work.

The ValueList::SetByLabel and Node::SetConfigParam were the only callers to ValueList::SetByValue(). So i changed the ValueList::SetByLabel to use the new SetByIndex().