Include new Hank DWS01 revision..
Hank released a revision of the DWS01 door/window sensor.  No changes from the original one configuration wise but the type ID was different so OpenZWave didn't detect the config params automatically.  Added an entry in manufacturer_specific.xml to point this new type id to the same config file as the old DWS01 and everything appears peachy.