Workaround for silently dropped frames (#1290)
In networks with heavy load, especially with lots of security frames (NONCE_GET) Controller Command could hang and never return.
The reason for this:
Receiving a NONCE_GET OZW handles this quite low level, below all queues. When a NONCE_GET is received, a NONCE_Report is immediately sent. If this NONCE_Report fails because of e.g. a CAN, then it is re-transmitted on queue level. But instead of a extra queue for this, the re-transmission gets done in Driver::WriteMsg(). The way this is done (as far as i understand it): If there is a nonce that needs to be re-transmitted, send it instead of the actual frame. The actual frame is then lost. While this is not that big of a deal for a SWITCH_BINARY_SET or something like this, if the "eaten" frame is e.g. a ControllerCommand_HasNodeFailed, then the Queued Controller Command (in the Controller Queue) will have the state 'starting and never recover from this.