Total text overhaul
Total overhaul of the labels and help texts except for the last item "Existence of endpoint 3 Action" (I'm unable to grasp what it's about) The original text seems to be fetched from the user manual and it kind of "cheesy" mildly put. It took me a while to actually understand what they meant.