Add Central Scene support for Aeotec WallMote Quad zw130
This adds scenes for buttons 1-4. It also adds configuration
parameters 9 & 10 so that the slide functions can be detected.

This was tested against the Dev branch and the master branch.
make xmltest passes

When you change a configuration parameter (i.e. enabling/disabling
the touch sound) you need to press the button on the back of the
wallmote in order for it to see it.

I also noticed that configuration 5 (the touch feedback color)
doesn't generate a config report, so the config file will always
say blue (value 0x0000ff00) even though the config set worked and
the LED color is as configured.