Change Interview Order to ensure we have the CC Version (#1876)
Originally, OZW would get the CC Version information much later in the interview cycle, but this was often too late for some classes that create ValueID's based on the version implemented by the device. The result was that after adding a device, the CC might be missing Values that are needed. This in particular was causing a crash in the SwitchMultiLevel, as we didn't create several Values during the creation of the CC, as we didn't have updated version information.

With this change, the interview orders have changed around a bit, and we also take the chance to do more interviewing of "static" information to save bootup times later on when reading from ozwcache.