adds node id to a controller state change notification
When receiving a controller state notification the node that a command was invoked on (if any) was always set to 0. In pyozw (python-openzwave) because there is no way to identify a node that was involved in a command when a change in the state of the controller takes place there was a lock added. this lock would stop any other commands from being issues if ozw has not sent a notification that the command has finished or it has failed. openzwave is designed to queue commands being sent. the problem is that if there is no way to identify the node that a  controller state update has been issued for if the same command has been issued on more then a single node there simply is no way of being able to track it.

A good example is if a node is sleeping... when the node awakens then the command gets issued. This could take a while to happen. during this time pyozw will not allow any other command to be sent. This lock is either going to cause the program to stall or the returned value is going to be False.. The issue with stalling is obvious.. the issue with returning False is that if ozw has an issue with any passed parameters it too will return False. An application has no way to know if False means there is a lock. or if False means there is an issue with data that is being passed. This lock does not need to be there at all if there is a way to identify the node that is involved in the controller state change. The mechanics in ozw have already been created to handle this. I believe it may have been an oversight. there are 2 other locations in the driver that set the node ID to 0, and I believe these are correct in setting them to 0.