Adds OpenZWave::Manager->GetNodeClassIds and OpenZWave::Node->GetCommandClassIds
The way it was done before.

Here is the setup.. you have 100 nodes on a network. the application
wants to be able to provide to controls depending on what command
classes they support.

so of the 50-60 or so application command classes in the ZWave protocol
(I know that OZW does not support all of them) in order to be able to
know what command classes a node supports you would have to call
OpenZWave::Manager->GetNodeClassInformation over and over again for
each of the command classes on each of the nodes.

so some simple math here 60 * 100 = 6000 function calls to determine
this really simple tidbit of information..

This new method takes the home id the node id and a pointer to an empty
vector<uint8>. If the method is successful it will return True and fill
the vector with the command class ids that the node supports. This
reduces the number of calls needed down to 100, the number of nodes.
Much more efficient this way. And also no need to create and fill
variables with information that is not going to get used.

I set up the method so that it will empty the vector before filling it.
So the same vector can get used between iteration loops over the nodes.