Adds GetValueMin/Max{value data type} methods.
I added separate data type gets for min and max for the values. The
original GetValueMin and GetValueMax are still in place. the new methods are

GetValueMinByte
GetValueMinShort
GetValueMinInt
GetValueMinDecimal

GetValueMaxByte
GetValueMaxShort
GetValueMaxInt
GetValueMaxDecimal

These methods return the following data types

uint_8: Byte
int16: Short
int32: Int
float: Decimal

The min and max are only written to and gotten from the config file for
ValueByte, ValueShort, ValueInt and ValueDecimal.

I believe that the user can evaluate the value in order to know which
method to call. I did add type conversion to int32 for GetValueMin and
GetValueMax. I am aware that I did not have to create the Byte, Short
and Int methods because int32 covers uint_8 and int16 also. I did it to
keep the flow for the API

From what I can from the ZWave specification only
COMMAND_CLASS_THERMOSTAT_SETPOINT version 3+ uses the decimal min and
max. I looked at the code for that command class and it appeared as tho
there were additional values that were created to handle the min and
max values. I think this was because there was no direct way to add
them other then what I have done. I did not remove these additional values.
I did set the max and min for the setpoint value so there really is no
need to have those other values in place.

This is a proof of concept and I am sure that changes will need to be made.
It compiles without error. I do not have a thermostat to be able to
test the GetValueMinDecimal and GetValueMaxDecimal. hopefully someone
that does will be willing to give it a go and see what happens.

I feel this is a change that needs to be made, as the ZWave
specification gets updated you will see more min and max values for
decimal specifically. I personally would love it if they would add onto
COMMAND_CLASS_SWITCH_MULTILEVEL and add fractions of a percent.
There are quite a few command classed that can be updated in this manner.