Create DZPA1
The Leviton DZ15S in-wall switch config was changed almost completely to the DZPA1 plug-in device, yet retained the DZ15S product type and zwave alliance "1957" ID so the DZ15S config didn't match either it or the DZPA1 device. 

This PR copies the DZ15S config to a new entry for DZPA1, changing the metadata that continued referencing the DZ15S to reference the equivalent DZPA1 information.  Note, I do not have a DZPA1 to test, but, I think this PR reflects the goal of modifying the DZ15S metadata probably so as to clone it to a new config for the DZPA1.  

I will next work to back out or otherwise change the DZ15S so its metadata will match its definition.