Added the European version of TZ67 called TZ67G
The European version of the TKB home TZ67 dimmer plug has a different ID and type compared to the included version. This version is called TZ67G

When copying the line for the TZ67 but changing the id and type accordingly, everything works perfectly. All functionality of the dimmer plug is there.